+++ src/afterfact.rs
@@ -217,8 +217,11 @@ fn emit_csv<W: std::io::Write>(
 ) -> io::Result<()> {
     let output_replaced_maps: HashMap<&str, &str> =
         HashMap::from_iter(vec![("🛂r", "\r"), ("🛂n", "\n"), ("🛂t", "\t")]);
-    let removed_replaced_maps: HashMap<&str, &str> =
+    let mut removed_replaced_maps: HashMap<&str, &str> =
         HashMap::from_iter(vec![("\n", " "), ("\r", " "), ("\t", " ")]);
+    if stored_static.multiline_flag {
+        removed_replaced_maps.insert(" ¦ ", "\r\n");
+    }
     let output_replacer = AhoCorasickBuilder::new()
         .match_kind(MatchKind::LeftmostLongest)
         .build(output_replaced_maps.keys());
@@ -391,16 +394,13 @@ fn emit_csv<W: std::io::Write>(
                     plus_header = false;
                 }
                 wtr.write_record(detect_info.ext_field.iter().map(|x| {
-                    output_remover
-                        .replace_all(
-                            &output_replacer.replace_all(
-                                &x.1.to_value(),
-                                &output_replaced_maps.values().collect_vec(),
-                            ),
-                            &removed_replaced_maps.values().collect_vec(),
-                        )
-                        .split_whitespace()
-                        .join(" ")
+                    output_remover.replace_all(
+                        &output_replacer.replace_all(
+                            &x.1.to_value(),
+                            &output_replaced_maps.values().collect_vec(),
+                        ),
+                        &removed_replaced_maps.values().collect_vec(),
+                    )
                 }))?;
             }
             // 各種集計作業
@@ -1682,6 +1682,7 @@ mod tests {
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv.csv").to_path_buf()),
+            multiline: false,
         });
         let dummy_config = Some(Config {
             action: Some(dummy_action),
@@ -1930,6 +1931,313 @@ mod tests {
             }
         };
         assert!(remove_file("./test_emit_csv.csv").is_ok());
+    }
+
+    #[test]
+    fn test_emit_csv_output_with_multiline_opt() {
+        let mock_ch_filter = message::create_output_filter_config(
+            "test_files/config/channel_abbreviations.txt",
+            true,
+        );
+        let test_filepath: &str = "test.evtx";
+        let test_rulepath: &str = "test-rule.yml";
+        let test_title = "test_title";
+        let test_level = "high";
+        let test_computername = "testcomputer";
+        let test_computername2 = "testcomputer2";
+        let test_eventid = "1111";
+        let test_channel = "Sec";
+        let output = "pokepoke";
+        let test_attack = "execution/txxxx.yyy";
+        let test_recinfo = "CommandRLine: hoge ¦ Test1: hogetest1 ¦ Test2: hogetest2";
+        let test_record_id = "11111";
+        let expect_time = Utc
+            .datetime_from_str("1996-02-27T01:05:01Z", "%Y-%m-%dT%H:%M:%SZ")
+            .unwrap();
+        let expect_tz = expect_time.with_timezone(&Utc);
+        let dummy_action = Action::CsvTimeline(CsvOutputOption {
+            output_options: OutputOption {
+                input_args: InputOption {
+                    directory: None,
+                    filepath: None,
+                    live_analysis: false,
+                },
+                profile: Some("verbose-2".to_string()),
+                enable_deprecated_rules: false,
+                exclude_status: None,
+                min_level: "informational".to_string(),
+                exact_level: None,
+                enable_noisy_rules: false,
+                end_timeline: None,
+                start_timeline: None,
+                eid_filter: false,
+                european_time: false,
+                iso_8601: false,
+                rfc_2822: false,
+                rfc_3339: false,
+                us_military_time: false,
+                us_time: false,
+                utc: false,
+                visualize_timeline: false,
+                rules: Path::new("./rules").to_path_buf(),
+                html_report: None,
+                no_summary: true,
+                common_options: CommonOptions {
+                    no_color: false,
+                    quiet: false,
+                },
+                detect_common_options: DetectCommonOption {
+                    evtx_file_ext: None,
+                    thread_number: None,
+                    quiet_errors: false,
+                    config: Path::new("./rules/config").to_path_buf(),
+                    verbose: false,
+                    json_input: false,
+                },
+                enable_unsupported_rules: false,
+            },
+            geo_ip: None,
+            output: Some(Path::new("./test_emit_csv_multiline.csv").to_path_buf()),
+            multiline: true,
+        });
+        let dummy_config = Some(Config {
+            action: Some(dummy_action),
+            debug: false,
+        });
+        let stored_static = StoredStatic::create_static_data(dummy_config);
+        let output_profile: Vec<(CompactString, Profile)> = load_profile(
+            "test_files/config/default_profile.yaml",
+            "test_files/config/profiles.yaml",
+            Some(&stored_static),
+        )
+        .unwrap_or_default();
+        {
+            let messages = &message::MESSAGES;
+            messages.clear();
+            let val = r##"
+                {
+                    "Event": {
+                        "EventData": {
+                            "CommandRLine": "hoge",
+                            "Test1": "hogetest1",
+                            "Test2": "hogetest2"
+                        },
+                        "System": {
+                            "TimeCreated_attributes": {
+                                "SystemTime": "1996-02-27T01:05:01Z"
+                            }
+                        }
+                    }
+                }
+            "##;
+            let event: Value = serde_json::from_str(val).unwrap();
+            let output_option = OutputOption {
+                input_args: InputOption {
+                    directory: None,
+                    filepath: None,
+                    live_analysis: false,
+                },
+                profile: Some("verbose-2".to_string()),
+                enable_deprecated_rules: false,
+                exclude_status: None,
+                min_level: "informational".to_string(),
+                exact_level: None,
+                enable_noisy_rules: false,
+                end_timeline: None,
+                start_timeline: None,
+                eid_filter: false,
+                european_time: false,
+                iso_8601: false,
+                rfc_2822: false,
+                rfc_3339: false,
+                us_military_time: false,
+                us_time: false,
+                utc: false,
+                visualize_timeline: false,
+                rules: Path::new("./rules").to_path_buf(),
+                html_report: None,
+                no_summary: false,
+                common_options: CommonOptions {
+                    no_color: false,
+                    quiet: false,
+                },
+                detect_common_options: DetectCommonOption {
+                    evtx_file_ext: None,
+                    thread_number: None,
+                    quiet_errors: false,
+                    config: Path::new("./rules/config").to_path_buf(),
+                    verbose: false,
+                    json_input: false,
+                },
+                enable_unsupported_rules: false,
+            };
+            let mut profile_converter: HashMap<&str, Profile> = HashMap::from([
+                (
+                    "Timestamp",
+                    Profile::Timestamp(format_time(&expect_time, false, &output_option)),
+                ),
+                (
+                    "Computer",
+                    Profile::Computer(CompactString::from(test_computername2)),
+                ),
+                (
+                    "Channel",
+                    Profile::Channel(
+                        mock_ch_filter
+                            .get(&CompactString::from("security"))
+                            .unwrap_or(&CompactString::default())
+                            .to_owned(),
+                    ),
+                ),
+                ("Level", Profile::Level(CompactString::from(test_level))),
+                (
+                    "EventID",
+                    Profile::EventID(CompactString::from(test_eventid)),
+                ),
+                (
+                    "MitreAttack",
+                    Profile::MitreTactics(CompactString::from(test_attack)),
+                ),
+                (
+                    "RecordID",
+                    Profile::RecordID(CompactString::from(test_record_id)),
+                ),
+                (
+                    "RuleTitle",
+                    Profile::RuleTitle(CompactString::from(test_title)),
+                ),
+                (
+                    "AllFieldInfo",
+                    Profile::AllFieldInfo(CompactString::from(test_recinfo)),
+                ),
+                (
+                    "RuleFile",
+                    Profile::RuleFile(CompactString::from(test_rulepath)),
+                ),
+                (
+                    "EvtxFile",
+                    Profile::EvtxFile(CompactString::from(test_filepath)),
+                ),
+                ("Tags", Profile::MitreTags(CompactString::from(test_attack))),
+            ]);
+            let eventkey_alias = load_eventkey_alias(
+                utils::check_setting_path(
+                    &CURRENT_EXE_PATH.to_path_buf(),
+                    "rules/config/eventkey_alias.txt",
+                    true,
+                )
+                .unwrap()
+                .to_str()
+                .unwrap(),
+            );
+            message::insert(
+                &event,
+                CompactString::new(output),
+                DetectInfo {
+                    rulepath: CompactString::from(test_rulepath),
+                    ruletitle: CompactString::from(test_title),
+                    level: CompactString::from(test_level),
+                    computername: CompactString::from(test_computername2),
+                    eventid: CompactString::from(test_eventid),
+                    detail: CompactString::default(),
+                    ext_field: output_profile.to_owned(),
+                    is_condition: false,
+                },
+                expect_time,
+                &mut profile_converter,
+                false,
+                &eventkey_alias,
+            );
+            *profile_converter.get_mut("Computer").unwrap() =
+                Profile::Computer(CompactString::from(test_computername));
+
+            message::insert(
+                &event,
+                CompactString::new(output),
+                DetectInfo {
+                    rulepath: CompactString::from(test_rulepath),
+                    ruletitle: CompactString::from(test_title),
+                    level: CompactString::from(test_level),
+                    computername: CompactString::from(test_computername),
+                    eventid: CompactString::from(test_eventid),
+                    detail: CompactString::default(),
+                    ext_field: output_profile.to_owned(),
+                    is_condition: false,
+                },
+                expect_time,
+                &mut profile_converter,
+                false,
+                &eventkey_alias,
+            );
+            let multi = message::MESSAGES.get(&expect_time).unwrap();
+            let (_, detect_infos) = multi.pair();
+
+            println!("message: {detect_infos:?}");
+        }
+        let expect =
+            "\"Timestamp\",\"Computer\",\"Channel\",\"EventID\",\"Level\",\"Tags\",\"RecordID\",\"RuleTitle\",\"Details\",\"AllFieldInfo\"\n\""
+                .to_string()
+                + &expect_tz.with_timezone(&Local).format("%Y-%m-%d %H:%M:%S%.3f %:z").to_string()
+                + "\",\""
+                + test_computername
+                + "\",\""
+                + test_channel
+                + "\","
+                + test_eventid
+                + ",\""
+                + test_level
+                + "\",\""
+                + test_attack
+                + "\","
+                + test_record_id
+                + ",\""
+                + test_title
+                + "\",\""
+                + output
+                + "\",\""
+                + &test_recinfo.replace(" ¦ ", "\r\n")
+                + "\"\n\""
+                + &expect_tz.with_timezone(&Local).format("%Y-%m-%d %H:%M:%S%.3f %:z")
+                .to_string()
+                + "\",\""
+                + test_computername2
+                + "\",\""
+                + test_channel
+                + "\","
+                + test_eventid
+                + ",\""
+                + test_level
+                + "\",\""
+                + test_attack
+                + "\","
+                + test_record_id
+                + ",\""
+                + test_title
+                + "\",\""
+                + output
+                + "\",\""
+                + &test_recinfo.replace(" ¦ ", "\r\n")
+                + "\"\n";
+        let mut file: Box<dyn io::Write> =
+            Box::new(File::create("./test_emit_csv_multiline.csv").unwrap());
+
+        assert!(emit_csv(
+            &mut file,
+            false,
+            HashMap::new(),
+            1,
+            &output_profile,
+            &stored_static,
+            (&Some(expect_tz), &Some(expect_tz))
+        )
+        .is_ok());
+        match read_to_string("./test_emit_csv_multiline.csv") {
+            Err(_) => panic!("Failed to open file."),
+            Ok(s) => {
+                assert_eq!(s, expect);
+            }
+        };
+        assert!(remove_file("./test_emit_csv_multiline.csv").is_ok());
     }
 
     #[test]