+++ src/detections/rule/condition_parser.rs
@@ -596,6 +596,7 @@ mod tests {
                 },
                 geo_ip: None,
                 output: None,
+                multiline: false,
             })),
             debug: false,
         }))