+++ src/options/profile.rs
@@ -443,6 +443,7 @@ mod tests {
                 },
                 geo_ip: None,
                 output: None,
+                multiline: false,
             }));
         assert_eq!(
             Nested::<Vec<String>>::new(),
@@ -551,6 +552,7 @@ mod tests {
                 },
                 geo_ip: None,
                 output: None,
+                multiline: false,
             }));
         *GEOIP_DB_PARSER.write().unwrap() = None;
         assert_eq!(
@@ -609,6 +611,7 @@ mod tests {
                 },
                 geo_ip: None,
                 output: None,
+                multiline: false,
             }));
 
         let expect: Vec<(CompactString, Profile)> = vec![
@@ -697,6 +700,7 @@ mod tests {
                 },
                 geo_ip: None,
                 output: None,
+                multiline: false,
             }));
         //両方のファイルが存在しない場合
         assert_eq!(