+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes
 
+## 2.3.2 [202X/XX/XX] "XXX"
+
+**Enhancements:**
+
+- Added `-M, --multiline` option in the `csv-timeline` command. (#972) (@hitenkoku)
+
 ## 2.3.1 [2023/03/18] "TMCIT Release-2"
 
 **Enhancements:**